requirements.txt
line-bot-sdk
flask
azure-cognitiveservices-vison-face
msrestazure